#665: Simplified UserMessage, loader and errors handling
