+++ AtlasUI/AtlasUI/AtlasUIViewController.swift
@@ -28,7 +28,7 @@ class AtlasUIViewController: UIViewController {
 
     override func viewDidLoad() {
         super.viewDidLoad()
-        UserMessage.loadBannerError()
+        UserError.loadBannerError()
         addChildViewController(mainNavigationController)
         view.addSubview(mainNavigationController.view)
         mainNavigationController.view.fillInSuperview()
@@ -50,6 +50,13 @@ extension AtlasUIViewController {
     func hideLoader() {
         loaderView.hide()
         loaderView.removeFromSuperview()
+    }
+
+    static func displayLoader(block: (@escaping () -> Void) -> Void) {
+        shared?.showLoader()
+        block {
+            shared?.hideLoader()
+        }
     }
 
 }
