+++ AtlasUI/AtlasUI/Utils/AtlasResult.swift
@@ -23,17 +23,13 @@ extension AtlasResult {
         }
     }
 
-    func process(forceFullScreenError fullScreen: Bool = false) -> T? {
+    func process(presentationMode: PresentationMode? = nil) -> T? {
         let processedResult = self.processedResult()
         switch processedResult {
         case .success(let data):
             return data
         case .error(let error, _, _):
-            if fullScreen {
-                UserMessage.displayErrorFullScreen(error: error)
-            } else {
-                UserMessage.displayError(error: error)
-            }
+            UserError.display(error: error, mode: presentationMode)
             return nil
         }
     }
