+++ AtlasUI/AtlasUI/Utils/AtlasUIClient.swift
@@ -8,7 +8,7 @@ import AtlasSDK
 struct AtlasUIClient {
 
     static func customer(completion: @escaping CustomerCompletion) {
-        UserMessage.displayLoader { hideLoader in
+        AtlasUIViewController.displayLoader { hideLoader in
             AtlasAPIClient.shared?.customer { result in
                 hideLoader()
                 completion(result)
@@ -17,7 +17,7 @@ struct AtlasUIClient {
     }
 
     static func createCart(cartItemRequests: [CartItemRequest], completion: @escaping CartCompletion) {
-        UserMessage.displayLoader { hideLoader in
+        AtlasUIViewController.displayLoader { hideLoader in
             AtlasAPIClient.shared?.createCart(withItems: cartItemRequests) { result in
                 hideLoader()
                 completion(result)
@@ -28,7 +28,7 @@ struct AtlasUIClient {
     static func createCheckoutCart(forSelectedArticle selectedArticle: SelectedArticle,
                                    addresses: CheckoutAddresses? = nil,
                                    completion: @escaping CheckoutCartCompletion) {
-        UserMessage.displayLoader { hideLoader in
+        AtlasUIViewController.displayLoader { hideLoader in
             AtlasAPIClient.shared?.createCheckoutCart(forSelectedArticle: selectedArticle, addresses: addresses) { result in
                 hideLoader()
                 completion(result)
@@ -37,7 +37,7 @@ struct AtlasUIClient {
     }
 
     static func createCheckout(cartId: String, addresses: CheckoutAddresses? = nil, completion: @escaping CheckoutCompletion) {
-        UserMessage.displayLoader { hideLoader in
+        AtlasUIViewController.displayLoader { hideLoader in
             AtlasAPIClient.shared?.createCheckout(fromCardId: cartId, addresses: addresses) { result in
                 hideLoader()
                 completion(result)
@@ -46,7 +46,7 @@ struct AtlasUIClient {
     }
 
     static func updateCheckout(checkoutId: String, updateCheckoutRequest: UpdateCheckoutRequest, completion: @escaping CheckoutCompletion) {
-        UserMessage.displayLoader { hideLoader in
+        AtlasUIViewController.displayLoader { hideLoader in
             AtlasAPIClient.shared?.updateCheckout(withId: checkoutId, updateCheckoutRequest: updateCheckoutRequest) { result in
                 hideLoader()
                 completion(result)
@@ -55,7 +55,7 @@ struct AtlasUIClient {
     }
 
     static func createOrder(checkoutId: String, completion: @escaping OrderCompletion) {
-        UserMessage.displayLoader { hideLoader in
+        AtlasUIViewController.displayLoader { hideLoader in
             AtlasAPIClient.shared?.createOrder(fromCheckoutId: checkoutId) { result in
                 hideLoader()
                 completion(result)
@@ -64,7 +64,7 @@ struct AtlasUIClient {
     }
 
     static func createGuestOrder(request: GuestOrderRequest, completion: @escaping GuestOrderCompletion) {
-        UserMessage.displayLoader { hideLoader in
+        AtlasUIViewController.displayLoader { hideLoader in
             AtlasAPIClient.shared?.createGuestOrder(request: request) { result in
                 hideLoader()
                 completion(result)
@@ -73,7 +73,7 @@ struct AtlasUIClient {
     }
 
     static func guestCheckoutPaymentSelectionURL(request: GuestPaymentSelectionRequest, completion: @escaping URLCompletion) {
-        UserMessage.displayLoader { hideLoader in
+        AtlasUIViewController.displayLoader { hideLoader in
             AtlasAPIClient.shared?.guestCheckoutPaymentSelectionURL(request: request) { result in
                 hideLoader()
                 completion(result)
@@ -82,7 +82,7 @@ struct AtlasUIClient {
     }
 
     static func guestCheckout(checkoutId: String, token: String, completion: @escaping GuestCheckoutCompletion) {
-        UserMessage.displayLoader { hideLoader in
+        AtlasUIViewController.displayLoader { hideLoader in
             AtlasAPIClient.shared?.guestCheckout(checkoutId: checkoutId, token: token) { result in
                 hideLoader()
                 completion(result)
@@ -91,7 +91,7 @@ struct AtlasUIClient {
     }
 
     static func article(withSKU sku: String, completion: @escaping ArticleCompletion) {
-        UserMessage.displayLoader { hideLoader in
+        AtlasUIViewController.displayLoader { hideLoader in
             AtlasAPIClient.shared?.article(withSKU: sku) { result in
                 hideLoader()
                 completion(result)
@@ -100,7 +100,7 @@ struct AtlasUIClient {
     }
 
     static func addresses(completion: @escaping AddressesCompletion) {
-        UserMessage.displayLoader { hideLoader in
+        AtlasUIViewController.displayLoader { hideLoader in
             AtlasAPIClient.shared?.addresses { result in
                 hideLoader()
                 completion(result)
@@ -109,7 +109,7 @@ struct AtlasUIClient {
     }
 
     static func deleteAddress(withId addressId: String, completion: @escaping SuccessCompletion) {
-        UserMessage.displayLoader { hideLoader in
+        AtlasUIViewController.displayLoader { hideLoader in
             AtlasAPIClient.shared?.deleteAddress(withId: addressId) { result in
                 hideLoader()
                 completion(result)
@@ -118,7 +118,7 @@ struct AtlasUIClient {
     }
 
     static func createAddress(_ request: CreateAddressRequest, completion: @escaping AddressChangeCompletion) {
-        UserMessage.displayLoader { hideLoader in
+        AtlasUIViewController.displayLoader { hideLoader in
             AtlasAPIClient.shared?.createAddress(request) { result in
                 hideLoader()
                 completion(result)
@@ -127,7 +127,7 @@ struct AtlasUIClient {
     }
 
     static func updateAddress(addressId: String, request: UpdateAddressRequest, completion: @escaping AddressChangeCompletion) {
-        UserMessage.displayLoader { hideLoader in
+        AtlasUIViewController.displayLoader { hideLoader in
             AtlasAPIClient.shared?.updateAddress(withId: addressId, request: request) { result in
                 hideLoader()
                 completion(result)
@@ -136,7 +136,7 @@ struct AtlasUIClient {
     }
 
     static func checkAddress(_ request: CheckAddressRequest, completion: @escaping AddressCheckCompletion) {
-        UserMessage.displayLoader { hideLoader in
+        AtlasUIViewController.displayLoader { hideLoader in
             AtlasAPIClient.shared?.checkAddress(request) { result in
                 hideLoader()
                 completion(result)
